Add `skipmissing` argument to `levels` (#46)
Currently if one wants to get levels in their appropriate order,
but add `missing` if present, the only solution is to do something like
`union(levels(x), unique(x))`, which is inefficient.
Support `skipmissing=false` to allow doing this in a single pass over the data.

Use `@inline` to ensure that the return type can be inferred when the value
of `skipmissing` is known statically. Also fix a type instability which existed
for ranges.

Use a complex strategy to ensure that the new argument is supported even
when a custom type defined `levels` without supporting it.